+++ src/frontend/js/components/CourseGlimpseList/CourseGlimpseList.spec.tsx
@@ -23,14 +23,36 @@ describe('components/CourseGlimpseList', () => {
         title: 'Course 45',
       },
     ] as Course[];
-    const { getByText } = render(
+    const { getAllByText, getByText } = render(
       <IntlProvider locale="en">
-        <CourseGlimpseList courses={courses} />
+        <CourseGlimpseList
+          courses={courses}
+          meta={{ limit: 20, offset: 0, total_count: 5 }}
+        />
       </IntlProvider>,
     );
 
+    expect(
+      getAllByText('Showing 5 courses matching your search').length,
+    ).toEqual(1);
     // Both courses' titles are shown
     getByText('Course 44');
     getByText('Course 45');
+  });
+
+  it('shows the count twice if there are more than 8 courses to show', () => {
+    const courses = [] as Course[];
+    const { getAllByText } = render(
+      <IntlProvider locale="en">
+        <CourseGlimpseList
+          courses={courses}
+          meta={{ limit: 20, offset: 0, total_count: 42 }}
+        />
+      </IntlProvider>,
+    );
+
+    expect(
+      getAllByText('Showing 42 courses matching your search').length,
+    ).toEqual(2);
   });
 });